@@ -3,8 +3,8 @@
   "version": "7.0.3",
   "description": "Extremely powerful dependency injection container.",
   "main": "lib/awilix.js",
-  "module": "lib/awilix.module.js",
-  "jsnext:main": "lib/awilix.module.js",
+  "module": "lib/awilix.module.mjs",
+  "jsnext:main": "lib/awilix.module.mjs",
   "browser": "lib/awilix.browser.js",
   "umd:main": "lib/awilix.umd.js",
   "react-native": "lib/awilix.browser.js",
@@ -50,37 +50,37 @@
   },
   "homepage": "https://github.com/jeffijoe/awilix#readme",
   "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/plugin-transform-runtime": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
-    "@babel/runtime": "^7.17.9",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.31",
-    "babel-jest": "^28.0.3",
+    "@babel/core": "^7.19.3",
+    "@babel/plugin-transform-runtime": "^7.19.1",
+    "@babel/preset-env": "^7.19.4",
+    "@babel/runtime": "^7.19.4",
+    "@types/jest": "^29.1.2",
+    "@types/node": "^18.8.5",
+    "babel-jest": "^29.2.0",
     "coveralls": "^3.1.1",
-    "husky": "^7.0.4",
+    "husky": "^8.0.1",
     "istanbul": "^0.4.5",
-    "jest": "^28.0.3",
-    "lint-staged": "^12.4.1",
-    "prettier": "^2.6.2",
+    "jest": "^29.2.0",
+    "lint-staged": "^13.0.3",
+    "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
-    "rollup": "^2.71.1",
+    "rollup": "^3.1.0",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-typescript2": "^0.31.2",
+    "rollup-plugin-typescript2": "^0.34.1",
     "smid": "^0.1.1",
-    "ts-jest": "^28.0.0-next.3",
+    "ts-jest": "^29.0.3",
     "tslib": "^2.4.0",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^9.0.0",
-    "typescript": "^4.6.4"
+    "typescript": "^4.8.4"
   },
   "dependencies": {
     "camel-case": "^4.1.2",
-    "fast-glob": "^3.2.11"
+    "fast-glob": "^3.2.12"
   },
   "lint-staged": {
     "*.ts": [
@@ -98,7 +98,7 @@
         "@babel/preset-env",
         {
           "targets": {
-            "node": "12.0.0"
+            "node": "14.0.0"
           }
         }
       ]
@@ -132,8 +132,13 @@
       "js"
     ],
     "transform": {
-      "^.+\\.tsx?$": "ts-jest",
-      "^.+\\.jsx?$": "babel-jest"
+      "^.+\\.tsx?$": [
+        "ts-jest",
+        {
+          "useESM": true
+        }
+      ],
+      "^.+\\.m?jsx?$": "babel-jest"
     }
   }
 }
